+++ src/components/login/background/background.css
@@ -1,0 +1,17 @@
+@keyframes move {
+  100% {
+    transform: translate3d(0, 0, 1px) rotate(360deg);
+  }
+}
+
+.background {
+  position: fixed;
+  width: 100vw;
+  height: 100vh;
+  top: 0;
+  left: 0;
+  background: #06071e;
+  overflow: hidden;
+  z-index: 1;
+}
+
