+++ src/components/login/background/backgroundCircle.tsx
@@ -1,0 +1,50 @@
+import React from 'react'
+import './backgroundCircle.css'
+
+function BackgroundCircle (): JSX.Element {
+  return (
+    <>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+      <span className='circle'></span>
+    </>
+  )
+}
+export default BackgroundCircle
