--- src/components/login/prueba/prueba.tsx
@@ -1,57 +1,0 @@
-import React from 'react'
-import './prueba.css'
-import Login from '../../../pages/Login'
-
-interface props {
-  children: JSX.Element
-}
-
-function Background ({ children }: props): JSX.Element {
-  return (
-    <>
-      <div className="background">
-        <Login />
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-        <span className='hola'></span>
-
-      </div>
-    </>
-  )
-}
-
-export default Background
